backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.9.0
anyio==4.0.0
async-timeout==4.0.3
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
envparse==0.2.0
fastapi==0.88.0
filelock==3.12.4
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.30
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
passlib==1.7.4
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pydantic==1.10.13
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.12.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.8.0
uvicorn==0.20.0
virtualenv==20.24.5
WTForms==3.0.1
yarl==1.9.2
requests==2.31.0
flower==2.0.1
celery==5.3.5
